// src/lib/sanity/sanity.types.ts
export type Seo = {
  _type: "seo";
  title?: string;
  description?: string;
  noindex?: boolean;
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type PageBySlugQueryResult = {
  _id: string;
  _updatedAt: string;
  title: string | null;
  slug: string | null;
  excerpt: string | null;
  body: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?:
          "blockquote" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        _key: string;
        _type: "callout";
        tone?: "info" | "warning";
        text?: string;
      }
    | {
        _key: string;
        _type: "codeBlock";
        language?: string;
        code?: string;
      }
    | {
        _key: string;
        _type: "imageWithAlt";
        asset: {
          _id: string;
          url: string | null;
          metadata: {
            dimensions: {
              width: number | null;
              height: number | null;
            } | null;
            lqip: string | null;
          } | null;
        } | null;
        media?: unknown;
        hotspot: SanityImageHotspot | null;
        crop: SanityImageCrop | null;
        alt: string | null;
        caption: string | null;
      }
  > | null;
  seo: Seo | null;
} | null;
// ...
export type PostBySlugQueryResult = {
  _id: string;
  _updatedAt: string;
  title: string | null;
  slug: string | null;
  excerpt: string | null;
  publishedAt: string | null;
  image: {
    _type: "imageWithAlt";
    alt: string | null;
    caption: string | null;
    hotspot: SanityImageHotspot | null;
    crop: SanityImageCrop | null;
    asset: {
      _id: string;
      url: string | null;
      metadata: {
        dimensions: {
          width: number | null;
          height: number | null;
        } | null;
        lqip: string | null;
      } | null;
    } | null;
  } | null;
  body: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?:
          "blockquote" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        _key: string;
        _type: "callout";
        tone?: "info" | "warning";
        text?: string;
      }
    | {
        _key: string;
        _type: "codeBlock";
        language?: string;
        code?: string;
      }
    | {
        _key: string;
        _type: "imageWithAlt";
        asset: {
          _id: string;
          url: string | null;
          metadata: {
            dimensions: {
              width: number | null;
              height: number | null;
            } | null;
            lqip: string | null;
          } | null;
        } | null;
        media?: unknown;
        hotspot: SanityImageHotspot | null;
        crop: SanityImageCrop | null;
        alt: string | null;
        caption: string | null;
      }
  > | null;
  seo: Seo | null;
} | null;